AlignedBuffer/Storage/ObjectStorage renamed to UninitializedBuffer/UninitializedStorage/GuardedUninitializedStorage
Added support for runtime buffer alignment (requires the compile time alignment to be set to 0). It's mostly useful when the implementation size is known, but can slightly differ (x86 vs x64 build for example) and it makes no sense trying to guess the exact size for every possible case. So you allocate a buffer for the largest size and align the object at the constructor call.